--- a/Resources/Prototypes/Entities/Structures/Walls/walls.yml
+++ b/Resources/Prototypes/Entities/Structures/Walls/walls.yml
@@ -1302,7 +1302,7 @@
   name: force wall
   components:
     - type: TimedDespawn
-      lifetime: 12
+      lifetime: 10
     - type: Tag
       tags:
         - Wall