<!-- Guidelines: https://docs.spacestation14.io/en/getting-started/pr-guideline -->

## About the PR
<!-- What did you change? -->

Force Wall was tweaked to be more fair and to prevent cheesing with wall layers to be invulnerable.

## Why / Balance
<!-- Discuss how this would affect game balance or explain why it was changed. Link any relevant discussions or issues. -->

Force Wall had some glaring issues:
- The despawn time was longer than the use action time
- If a wizard hid in the wall hitbox they couldn't be hit (they also couldn't interact out of it either)

Now the despawn timer is shorter than the use action time
12s => 10s
And the Action Use Delay is longer:
10s => 15s

A Special Wall Layer was added to compensate for this. It's almost the same as a Wall Layer except Impassible and InteractImpassible are removed from it. 
- Non-casters still can't go past it (including mice and flying creatures)
- It can't be shot through with bullets, or lasers, or any projectile from the wizard
- The wizard now CAN interact out of it, but they can also be hit while inside of it too

## Technical details
<!-- Summary of code changes for easier review. -->

SpecialWallLayer was added to collision groups, it's similar to a Wall Layer except that anyone inside of it can be interacted with and can interact outside of it.

## Media
<!-- Attach media if the PR makes ingame changes (clothing, items, features, etc). 
Small fixes/refactors are exempt. Media may be used in SS14 progress reports with credit. -->

https://github.com/user-attachments/assets/d5aa226e-5d03-4fcc-849d-c8c599ec9878

## Requirements
<!-- Confirm the following by placing an X in the brackets [X]: -->
- [x] I have read and am following the [Pull Request and Changelog Guidelines](https://docs.spacestation14.com/en/general-development/codebase-info/pull-request-guidelines.html).
- [x] I have added media to this PR or it does not require an ingame showcase.
<!-- You should understand that not following the above may get your PR closed at maintainer’s discretion -->

## Breaking changes
<!-- List any breaking changes, including namespaces, public class/method/field changes, prototype renames; and provide instructions for fixing them.
This will be posted in #codebase-changes. -->

**Changelog**
<!-- Add a Changelog entry to make players aware of new features or changes that could affect gameplay.
Make sure to read the guidelines and take this Changelog template out of the comment block in order for it to show up.
Changelog must have a :cl: symbol, so the bot recognizes the changes and adds them to the game's changelog. -->
<!--
-->
:cl:
- tweak: Force Wall timers changed so it despawns faster than it can be cast.
- tweak: Force Wall users can now interact while inside of the wall, but also can be attacked while inside of the wall.